backend/requirements.txt
# Backend — API + RAG (LangChain) + Supabase pgvector
fastapi>=0.115.0,<1
uvicorn[standard]>=0.32.0,<1
httpx>=0.27.0,<1
pytest>=8.0.0,<9
pydantic-settings>=2.0.0,<3
python-dotenv>=1.0.0,<2
supabase>=2.0.0,<3
langchain-core>=0.3.0
langchain-community>=0.3.0
langchain-openai>=0.2.0
langchain-huggingface>=0.1.0
sentence-transformers>=3.0.0
